requirements.txt
Django>=6.0
psycopg2-binary>=2.9
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv